fix: move PWL exception types into core/bess/exceptions.py
docs/agents/rules.md requires exception types live in exceptions.py
and nowhere else; PWLWindowUnderRefinedError and PWLEndSoeOutOfRangeError
were defined in pwl_window_dp.py instead.
